// src/_generated/UsernameExistenceMutation.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: UsernameExistenceMutation
// ====================================================

export interface UsernameExistenceMutation {
  checkUsernameExistence: boolean;
}

export interface UsernameExistenceMutationVariables {
  username: string;
}